--- a/scripts/bundle.ts
+++ b/scripts/bundle.ts
@@ -1,9 +1,9 @@
 import * as flags from "./deps/std_flags.ts";
 import { copy as fsCopy } from "./deps/std_fs_copy.ts";
-import { writeAll } from "./deps/std_io_util.ts";
 import * as path from "./deps/std_path.ts";
 
 import * as esbuild from "./deps/esbuild.ts";
+import { colorizeConsole } from "./deps/colorize.ts";
 
 // It doesn't have to be the same exact Ajv instance because the code
 // is stored on the `source` property of the Function object,
@@ -12,6 +12,8 @@ import ajv from "../src/schemas/_ajv_jtd.ts";
 import standaloneCode from "../src/deps/ajv_standalone.ts";
 
 import * as schemas from "../src/schemas/mod.ts";
+
+colorizeConsole();
 
 const { help: helpArg } = flags.parse(Deno.args, {
 	boolean: ["help"],
@@ -77,22 +79,16 @@ console.debug(`${repo(outFile)} is ${oldSize}B`);
 console.debug(`Copying ${repo(schemasDir)} to ${repo(buildDir)}`);
 await fsCopy(schemasDir, buildDir, { overwrite: true });
 
-const te = new TextEncoder();
-
-console.info("Compiling and writing route validators");
+console.info("Writing route validators");
 const timeWritingValidators = timer();
 await Promise.all([
 	...Object.entries(schemas.validators).map(
-		// Originally, I thought I had to transpile the validator to ESM, since AJV
-		// compiles validators to CJS. As it turns out, you can actually import CJS
-		// from ESM and esbuild will just... handle it. Somehow, it actually ended
-		// up slimming down the bundle by a few kB too. Dunno how, but that's cool!
 		async ([key, validator]) => {
-			console.debug(`Compiling validator: ${key}`);
+			console.debug(`Reading validator: ${key}`);
 			const src = standaloneCode(ajv, validator);
 
-			// There is a direct* map between the object key and file path.
-			// * so long as I name the routes and files consistently.
+			// There is a direct map between the object key and file path,
+			// so long as the files are named consistently.
 			const outFileName = key
 				.replaceAll(" ", "")
 				.replaceAll(":", "")
@@ -101,34 +97,25 @@ await Promise.all([
 			// Since the file will be transpiled during bundling and it won't be
 			// type checked at any stage of that, it's safe to write this CJS
 			// straight to a TS file. This also means I don't have to write a plugin
-			// to rewrite the imports of mod.ts.
+			// to rewrite the imports of mod.ts. esbuild will happily import CJS
+			// from an ESM file.
 			const outFilePath = path.join(buildDir, outFileName);
 
-			console.debug(`Opening ${repo(outFilePath)}`);
-			// Note the omission of `create` - if there's no file to open, that's an
-			// error, since `key` should directly map to a file.
-			const file = await Deno.open(outFilePath, {
-				write: true,
-				truncate: true,
-			});
-			const bytes = te.encode(src);
-
 			console.debug(`Writing code for '${key}' to ${repo(outFilePath)}`);
-			await writeAll(file, bytes);
-
-			console.debug(`Closing file ${repo(outFilePath)}`);
-			Deno.close(file.rid);
+			await Deno.writeTextFile(outFilePath, src, { create: false });
+			console.debug(`Finished writing ${repo(outFilePath)}`);
 		},
 	),
 ]);
-console.debug(`Written validators in ${timeWritingValidators()} ms`);
+console.debug(`Wrote validators in ${timeWritingValidators()} ms`);
 
 // Referencing ./mod.ts basically makes TS treat the bundled file as if it
 // were actually mod.ts. I have no idea if this is the intended behaviour, but
 // it works. It also has the super fun side-effect of crashing the Deno LSP
 // whenever you open the file in-editor, which makes development more exciting.
-// There are tests to ensure that mod.ts and mod.bundle.js behave identically.
-const banner = `/// <reference types="./mod.ts" />
+// Tests verify that schemas/mod.ts and schemas/mod.bundle.js behave identically.
+const banner = `
+/// <reference types="./mod.ts" />
 
 // dprint-ignore-file
 // deno-lint-ignore-file
@@ -139,7 +126,9 @@ const banner = `/// <reference types="./mod.ts" />
 // WARNING: opening this file soft-crashes the Deno language server. If you
 // opened this file in Visual Studio Code and use the vscode_deno extension,
 // restart or reload your editor now (it's faster than waiting for the LS)
-`;
+`.trimStart();
+
+const timeBundling = timer();
 
 console.info(`Bundling, write to ${repo(outFile)}`);
 await esbuild.build({
@@ -150,10 +139,14 @@ await esbuild.build({
 	banner: {
 		js: banner,
 	},
+	// Keeps stack traces clean and readable
 	minifyIdentifiers: false,
+	// With whitespace un-minified, GitHub reports that the majority of this
+	// repository is JavaScript because the bundle is so large :(
 	minifySyntax: true,
 	minifyWhitespace: true,
 });
+console.debug(`Finished bundling in ${timeBundling()} ms`);
 
 // Unlike in Node, esbuild requires manually calling stop() in Deno, otherwise
 // the program will keep running indefinitely.