--- a/scripts/deps/colorize.ts
+++ b/scripts/deps/colorize.ts
@@ -16,3 +16,30 @@ export const colorize = tagApply<[string, ...TextStyle[]], string>(
 			? names.reduce((str, color) => colors[color](str), string)
 			: string,
 );
+
+const { debug, log, info, warn, error } = console;
+
+/** Adds prefixes to the console logging functions. */
+export function colorizeConsole() {
+	const [preDebug, preLog, preInfo, preWarn, preError] = colorize`
+${["Debug", "magenta"]}
+${["Log", "cyan"]}
+${["Info", "yellow"]}
+${["Warn", "red"]}
+${[" Error ", "bgRed"]}
+`.trim().split("\n").map((line) => line.trimStart());
+
+	console.debug = debug.bind(console, preDebug);
+	console.log = log.bind(console, preLog);
+	console.info = info.bind(console, preInfo);
+	console.warn = warn.bind(console, preWarn);
+	console.error = error.bind(console, preError);
+}
+
+export function restoreConsole() {
+	console.debug = debug;
+	console.info = info;
+	console.log = log;
+	console.warn = warn;
+	console.error = error;
+}