--- a/scripts/deps/esbuild.ts
+++ b/scripts/deps/esbuild.ts
@@ -1,1 +1,1 @@
-export * from "https://deno.land/x/esbuild@v0.12.15/mod.js";
+export * from "https://deno.land/x/esbuild@v0.12.23/mod.js";