--- a/src/schemas/GET_devices.ts
+++ b/src/schemas/GET_devices.ts
@@ -203,4 +203,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);