--- a/src/schemas/GET_devices_groups.ts
+++ b/src/schemas/GET_devices_groups.ts
@@ -15,4 +15,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);