--- a/src/schemas/GET_devices_udid.ts
+++ b/src/schemas/GET_devices_udid.ts
@@ -116,4 +116,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);