--- a/src/schemas/GET_users.ts
+++ b/src/schemas/GET_users.ts
@@ -17,4 +17,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);