--- a/src/schemas/GET_users_groups.ts
+++ b/src/schemas/GET_users_groups.ts
@@ -15,4 +15,4 @@ const responseDataSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseDataSchema);
+export default ajv.compile(responseDataSchema);