--- a/src/schemas/GET_users_groups_id.ts
+++ b/src/schemas/GET_users_groups_id.ts
@@ -13,4 +13,4 @@ const responseDataSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseDataSchema);
+export default ajv.compile(responseDataSchema);