--- a/src/schemas/GET_users_id.ts
+++ b/src/schemas/GET_users_id.ts
@@ -13,4 +13,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);