--- a/src/schemas/POST_devices_udid_restart.ts
+++ b/src/schemas/POST_devices_udid_restart.ts
@@ -14,4 +14,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);