--- a/src/schemas/POST_devices_udid_wipe.ts
+++ b/src/schemas/POST_devices_udid_wipe.ts
@@ -12,4 +12,4 @@ const responseSchema: JTDSchemaType<ResponseData> = {
 	},
 };
 
-export default ajv.compileParser(responseSchema);
+export default ajv.compile(responseSchema);