--- a/src/schemas/_ajv_jtd.ts
+++ b/src/schemas/_ajv_jtd.ts
@@ -9,6 +9,7 @@ export * from "../deps/ajv_jtd.ts";
 /** Singleton instance of `Ajv` */
 export default new Ajv({
 	allErrors: true,
+	messages: true,
 	code: {
 		source: true,
 	},