--- a/test/api/getDevices.test.ts
+++ b/test/api/getDevices.test.ts
@@ -1,5 +1,5 @@
 import * as jamf from "../../src/mod.ts";
-import parseGetDevices from "../../src/schemas/GET_devices.ts";
+import validateGetDevices from "../../src/schemas/GET_devices.ts";
 import { assertEquals, assertThrowsAsync } from "../deps/asserts.ts";
 import { relativeTextFileReader } from "../deps/read_relative_file.ts";
 import * as mockFetch from "../deps/mock_fetch.ts";
@@ -15,7 +15,8 @@ const api = jamf.createAPI({
 mockFetch.install(); // we don't need to uninstall, all fetches should be mocked
 
 const jsonString = await readRelativeTextFile("../example_data/GET_devices__200.json");
-const jsonObject = parseGetDevices(jsonString) ?? (() => {
+const jsonObject = JSON.parse(jsonString);
+validateGetDevices(jsonObject) ?? (() => {
 	throw Error("Invalid data: ../example_data/GET_devices__200.json");
 })();