Fixes #1 

---

<details>
<summary><b>Parse & validate benchmarks</b></summary>
<br>

On main:

```console
$ deno run --allow-read --no-check .\test\schemas\bench_devices.ts
running 2 benchmarks ...
benchmark GET /devices ... 
    1000 runs avg: 16.338ms
benchmark GET /devices (with apps) ... 
    1000 runs avg: 56.646ms
benchmark result: DONE. 2 measured; 0 filtered
```

In this branch:

```console
$ deno run --allow-read --no-check .\test\schemas\bench_devices.ts
running 2 benchmarks ...
benchmark GET /devices ... 
    1000 runs avg: 0.858ms
benchmark GET /devices (with apps) ... 
    1000 runs avg: 3.88ms
benchmark result: DONE. 2 measured; 0 filtered
```

My assumption was that JSON.parse was well optimized, but I also suspected that there might be a cache that this is hitting. Here's data from a cold start, which is mostly how you'll be using the schemas anyway (you probably won't be validating the same data 1000 times in the same run). Basically, any caching is negligible, V8 is just magic.

Route: GET /devices
Method | Average time (ms)
--- | ---
AJV Validator, JSON.parse | 7.4
AJV Parser | 54.4

AJV's claim of "performance comparable to JSON.parse" is... misleading. Possibly true for smaller schemas and less data, but Jamf School's API definitely doesn't suit it well.

</details>

---

### Why bundle the validators?
AJV takes its sweet time to compile them, and [not all platforms support code generation](https://deno.com/deploy/docs/runtime-api#limitations).

Because compilation has to happen synchronously, import times _seriously_ suffer unless you do the work ahead of time. AJV allows you to export the compiled validator as CJS. esbuild can just take that and pretend it's ESM - no transpiling needed.

---

<details>
<summary><b>Import time benchmarks</b></summary>
<br>

On main:

```typescript
// import_parsers.ts
import "./src/schemas/mod.ts";
```
```console
$ hyperfine --warmup 3 'deno run --no-check import_parsers.ts'
Benchmark #1: deno run --no-check import_parsers.ts
  Time (mean ± σ):     274.9 ms ±   3.0 ms    [User: 434.8 ms, System: 18.2 ms]
  Range (min … max):   268.9 ms … 277.8 ms    10 runs
```

With this patch (comparing bundled, unbundled):

```typescript
// import_mod.ts
import "./src/schemas/mod.ts";
```
```typescript
// import_bundle.ts
import "./src/schemas/mod.bundle.js";
```
```console
$ hyperfine --warmup 3 'deno run --no-check import_bundle.ts' 'deno run --no-check import_mod.ts'
 
Benchmark #1: deno run --no-check import_bundle.ts
  Time (mean ± σ):      49.9 ms ±   0.4 ms    [User: 43.7 ms, System: 6.0 ms]
  Range (min … max):    49.2 ms …  50.9 ms    57 runs
 
Benchmark #2: deno run --no-check import_mod.ts
  Time (mean ± σ):     276.4 ms ±   3.2 ms    [User: 441.8 ms, System: 18.0 ms]
  Range (min … max):   270.5 ms … 279.9 ms    10 runs
 
Summary
  'deno run --no-check import_bundle.ts' ran
    5.53 ± 0.08 times faster than 'deno run --no-check import_mod.ts'
```

This also can't be benched using Deno's std/testing/bench.ts & dynamic imports, since even if you import schemas/mod.ts with a different URL (eg. adding ?1, ?2, etc) its dependencies are still cached. Import time benchmarks are only accurate from a cold start.

</details>

---

### Scripts

This PR adds 2 scripts: `scripts/bundle.ts` and `scripts/check_bundle.ts`
The check script will run in CI on each commit to ensure any changes to the schemas have been bundled.

The output is whitespace-minified because, in my pre-public experiments, GitHub was saying that 99% of this repository was JavaScript instead of Typescript. Yes, it's that petty :)