@SeparateRecords 2021-08-28
This needs more work on error messages. I haven't done any testing with them yet.